Improve baremetal deploy script.
Add ping.py so we can test when a host is up.  ping -c 2 -w 90 will
almost do the right thing, but will exit on a icmp host unreachable.
We need to continue to wait.  ping.py just waits for an answer to a
ping.

Parallelize building the compute nodes (but not the head node).